# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Maktab Raqamli Tizimi API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List users
      parameters:
        - in: query
          name: role
          schema:
            type: string
            enum: [director, zavuch, teacher, student, parent]
        - in: query
          name: classId
          schema:
            type: number
        - in: query
          name: search
          schema:
            type: string
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: number
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete user
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
...
  /subjects:
    get:
      operationId: listSubjects
      tags: [subjects]
      summary: List subjects
      responses:
        "200":
          description: List of subjects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subject"
    post:
      operationId: createSubject
      tags: [subjects]
      summary: Create subject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubjectInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subject"
  /subjects/{id}:
    get:
      operationId: getSubject
      tags: [subjects]
      summary: Get subject
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Subject
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subject"
    patch:
      operationId: updateSubject
      tags: [subjects]
      summary: Update subject
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubjectUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subject"
    delete:
      operationId: deleteSubject
      tags: [subjects]
      summary: Delete subject
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
...
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
  /messages:
    get:
      operationId: listMessages
      tags: [messages]
      summary: List messages
      parameters:
        - in: query
          name: withUserId
          schema:
            type: number
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
...
components:
  schemas:
...
    User:
      type: object
      required: [id, firstName, lastName, role, username, createdAt]
      properties:
        id:
          type: number
        firstName:
          type: string
        lastName:
          type: string
        username:
          type: string
        role:
          type: string
          enum: [director, zavuch, teacher, student, parent]
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        classId:
          type: ["integer", "null"]
        parentId:
          type: ["integer", "null"]
        subjectIds:
          type: array
          items:
            type: number
        createdAt:
          type: string
          format: date-time
    UserInput:
      type: object
      required: [firstName, lastName, username, password, role]
      properties:
        firstName:
          type: string
        lastName:
          type: string
        username:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [director, zavuch, teacher, student, parent]
        email:
          type: string
        phone:
          type: string
        classId:
          type: number
        parentId:
          type: number
    UserUpdate:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        classId:
          type: number
...
    Subject:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: number
        name:
          type: string
        description:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    SubjectInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
    SubjectUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
...
    Notification:
      type: object
      required: [id, userId, type, title, isRead, createdAt]
      properties:
        id:
          type: number
        userId:
          type: number
        type:
          type: string
          enum: [grade, attendance, announcement, message, warning, assignment]
        title:
          type: string
        body:
          type: ["string", "null"]
        isRead:
          type: boolean
        relatedId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
    Message:
      type: object
      required: [id, senderId, receiverId, content, createdAt]
      properties:
        id:
          type: number
        senderId:
          type: number
        senderName:
          type: string
        receiverId:
          type: number
        receiverName:
          type: string
        content:
          type: string
        isRead:
          type: boolean
        createdAt:
          type: string
          format: date-time
    MessageInput:
      type: object
      required: [receiverId, content]
      properties:
        receiverId:
          type: number
        content:
          type: string
...