# lib/api-zod/src/generated/types/subjectUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Maktab Raqamli Tizimi API
 * OpenAPI spec version: 0.1.0
 */

export interface SubjectUpdate {
  name?: string;
  description?: string;
}